Make YAFluxRegister template
Make YAFluxRegister a template so that it can be used for single-precision
data when AMReX is built for double precision.  The class is now called
YAFluxRegisterT and YAFluxRegister is aliased to YAFluxRegisterT<MultiFab>.
So there are no changes in the user's code.
