+++ Src/Boundary/AMReX_YAFluxRegister_2D_K.H
@@ -2,14 +2,15 @@
 #define AMREX_YAFLUXREGISTER_2D_K_H_
 #include <AMReX_Config.H>
 
-#include <AMReX_FArrayBox.H>
+#include <AMReX_BaseFab.H>
 
 namespace amrex {
 
+template <typename T>
 AMREX_GPU_HOST_DEVICE AMREX_FORCE_INLINE
-void yafluxreg_crseadd (Box const& bx, Array4<Real> const& d, Array4<int const> const& flag,
-                        Array4<Real const> const& fx, Array4<Real const> const& fy,
-                        Real dtdx, Real dtdy, int nc) noexcept
+void yafluxreg_crseadd (Box const& bx, Array4<T> const& d, Array4<int const> const& flag,
+                        Array4<T const> const& fx, Array4<T const> const& fy,
+                        T dtdx, T dtdy, int nc) noexcept
 {
     auto const lo = amrex::lbound(bx);
     auto const hi = amrex::ubound(bx);
@@ -43,9 +44,10 @@ void yafluxreg_crseadd (Box const& bx, Array4<Real> const& d, Array4<int const> 
     }}
 }
 
+template <typename T>
 AMREX_GPU_HOST_DEVICE AMREX_FORCE_INLINE
-void yafluxreg_fineadd (Box const& bx, Array4<Real> const& d, Array4<Real const> const& f,
-                        Real dtdx, int nc, int dirside, Dim3 const& rr) noexcept
+void yafluxreg_fineadd (Box const& bx, Array4<T> const& d, Array4<T const> const& f,
+                        T dtdx, int nc, int dirside, Dim3 const& rr) noexcept
 {
     const auto lo = amrex::lbound(bx);
     const auto hi = amrex::ubound(bx);
@@ -57,10 +59,10 @@ void yafluxreg_fineadd (Box const& bx, Array4<Real> const& d, Array4<Real const>
             for (int j = lo.y; j <= hi.y; ++j) {
                 const int i = lo.x;
                 const int ii = (i+1)*rr.x;
-                Real* AMREX_RESTRICT dp = &(d(i,j,0,n));
+                T* AMREX_RESTRICT dp = &(d(i,j,0,n));
                 for (int joff = 0; joff < rr.y; ++joff) {
                     const int jj =      j*rr.y  + joff;
-                    Real tmp = -dtdx*f(ii,jj,0,n);
+                    T tmp = -dtdx*f(ii,jj,0,n);
                     HostDevice::Atomic::Add(dp, tmp);
                 }
             }
@@ -73,10 +75,10 @@ void yafluxreg_fineadd (Box const& bx, Array4<Real> const& d, Array4<Real const>
             for (int j = lo.y; j <= hi.y; ++j) {
                 const int i = lo.x;
                 const int ii = i*rr.x;
-                Real* AMREX_RESTRICT dp = &(d(i,j,0,n));
+                T* AMREX_RESTRICT dp = &(d(i,j,0,n));
                 for (int joff = 0; joff < rr.y; ++joff) {
                     const int jj =      j*rr.y  + joff;
-                    Real tmp = dtdx*f(ii,jj,0,n);
+                    T tmp = dtdx*f(ii,jj,0,n);
                     HostDevice::Atomic::Add(dp, tmp);
                 }
             }
@@ -89,10 +91,10 @@ void yafluxreg_fineadd (Box const& bx, Array4<Real> const& d, Array4<Real const>
             for (int i = lo.x; i <= hi.x; ++i) {
                 const int j = lo.y;
                 const int jj = (j+1)*rr.y;
-                Real* AMREX_RESTRICT dp = &(d(i,j,0,n));
+                T* AMREX_RESTRICT dp = &(d(i,j,0,n));
                 for (int ioff = 0; ioff < rr.x; ++ioff) {
                     const int ii =      i*rr.x  + ioff;
-                    Real tmp = -dtdx*f(ii,jj,0,n);
+                    T tmp = -dtdx*f(ii,jj,0,n);
                     HostDevice::Atomic::Add(dp, tmp);
                 }
             }
@@ -105,10 +107,10 @@ void yafluxreg_fineadd (Box const& bx, Array4<Real> const& d, Array4<Real const>
             for (int i = lo.x; i <= hi.x; ++i) {
                 const int j = lo.y;
                 const int jj = j*rr.y;
-                Real* AMREX_RESTRICT dp = &(d(i,j,0,n));
+                T* AMREX_RESTRICT dp = &(d(i,j,0,n));
                 for (int ioff = 0; ioff < rr.x; ++ioff) {
                     const int ii =      i*rr.x  + ioff;
-                    Real tmp = dtdx*f(ii,jj,0,n);
+                    T tmp = dtdx*f(ii,jj,0,n);
                     HostDevice::Atomic::Add(dp, tmp);
                 }
             }
